Experimental attempt at using Document UI
This uses a document group to handle opening not just one, but multiple files!

Big questions are saving (that is, it shouldn't have it... except for playlists. SwiftUI bug...) and playlists (how will we handle this?).

This also encapsulates module specific state into a ModuleState class. Even if this doesn't work out, extracting this refractor would make a lot of sense.